refactor: split oversized processing modules
