+++ src/browser/app-state.ts
@@ -1,0 +1,18 @@
+import type { RGB } from "../shared/types";
+
+export type ProcessingState = {
+	currentFixedPalette?: RGB[];
+	currentExtractedPalette: RGB[];
+	compareBeforeOriginalUrl: string;
+	compareBeforeSanitizedUrl: string;
+	compareAfterUrl: string;
+	compareBeforeMode: "original" | "sanitized";
+};
+
+export const createProcessingState = (): ProcessingState => ({
+	currentExtractedPalette: [],
+	compareBeforeOriginalUrl: "",
+	compareBeforeSanitizedUrl: "",
+	compareAfterUrl: "",
+	compareBeforeMode: "original",
+});
